pyproject.toml
[tool.poetry]
name = "ezt"
version = "0.1.0"
description = ""
authors = ["John Kaustinen <jokausti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.3"
Jinja2 = "^3.0.3"
PyYAML = "^6.0"
datafusion = "^0.5.2"
polars = "^0.13.38"
s3fs = "^2021.11.1"
pyparsing = "3.0.6"
deltalake = "^0.6.2"
graphlib-backport = "^1.0.3"
networkx = "^2.8"
matplotlib = "^3.5.1"
rich = "^12.2.0"
rich-click = "^1.3.0"
jsonschema = "^4.14.0"
pyarrow = "^9.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
dprint = "^0.1.0"
types-PyYAML = "^6.0.7"
coverage = "^6.4.4"
pytest-mock = "^3.8.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ezt = "ezt.main_cli:ezt"